chore: separate form validation logic from context
